add demo3 多列唯一索引的用例

+++ cdc/src/main/java/com/sauzny/cdc/unique/UniqueLock.java
@@ -9,9 +9,9 @@ public class UniqueLock {
 
     private int wait_count;
 
-    private int event_id;
+    private long event_id;
 
-    private Map<Integer, Boolean> cond_map = new HashMap<>();
+    private Map<Long, Boolean> cond_map = new HashMap<>();
 
     public int getWait_count() {
         return wait_count;
@@ -21,19 +21,19 @@ public class UniqueLock {
         this.wait_count = wait_count;
     }
 
-    public int getEvent_id() {
+    public long getEvent_id() {
         return event_id;
     }
 
-    public void setEvent_id(int event_id) {
+    public void setEvent_id(long event_id) {
         this.event_id = event_id;
     }
 
-    public Map<Integer, Boolean> getCond_map() {
+    public Map<Long, Boolean> getCond_map() {
         return cond_map;
     }
 
-    public void setCond_map(Map<Integer, Boolean> cond_map) {
+    public void setCond_map(Map<Long, Boolean> cond_map) {
         this.cond_map = cond_map;
     }
